storage: move 'Modify storage' to the header kebab
Merge CockpitStorageIntegration and ModifyStorage files into one, having
the "Modify storage" button split from the cockpit integration page implementation
was slightly weird.

Also stop utilizing the 'isFormDisabled' in the
CockpitStorageIntegration, we don't disable the form anyhow from within
cockpit-storage mode, so this was causing unnecessary complexity in the
code.
